# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: MoMo APIs
  description: HTTP Service for accessing MTN's MoMo API.
  termsOfService: https://github.com/wondenge/momo-go/blob/master/LICENCE
  contact:
    name: William Ondenge
    email: ondengew@gmail.com
    url: https://www.ondenge.me
  license:
    name: Apache License
    url: https://github.com/wondenge/momo-go/blob/master/LICENCE
  version: "1.0"
host: sandbox.momodeveloper.mtn.com
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /apiuser/{APIUser}:
    get:
      tags:
      - user
      summary: GetUserDetails user
      description: GET API User Details
      operationId: user#GetUserDetails
      produces:
      - application/json
      parameters:
      - name: APIUser
        in: path
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserGetUserDetailsResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserGetUserDetailsBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserGetUserDetailsInternalErrorResponseBody'
      schemes:
      - https
...
  /v1_0/apiuser:
    post:
      tags:
      - user
      summary: NewUser user
      description: Used to create an API user in the sandbox target environment.
      operationId: user#NewUser
      parameters:
      - name: NewUserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserNewUserRequestBody'
      responses:
        "201":
          description: Created response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserNewUserBadRequestResponseBody'
        "409":
          description: Conflict response.
          schema:
            $ref: '#/definitions/UserNewUserConflictResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserNewUserInternalErrorResponseBody'
      schemes:
      - https
  /v1_0/apiuser/{X-Reference-Id}:
    get:
      tags:
      - user
      summary: GetUser user
      description: Used to get API user information.
      operationId: user#GetUser
      parameters:
      - name: GetUserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserGetUserRequestBody'
          required:
          - X-Reference-Id
      responses:
        "200":
          description: OK response.
          schema:
            type: string
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UserGetUserBadRequestResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/UserGetUserNotFoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/UserGetUserInternalErrorResponseBody'
      schemes:
      - https
...
definitions:
...
  PaymentServerUrlResponseBody:
    title: PaymentServerUrlResponseBody
    type: object
    properties:
      apiKey:
        type: string
        description: The payment server URL
        example: Rerum assumenda dolor.
    example:
      apiKey: Dolorem quo deleniti velit ratione esse doloribus.
...
  TargetEnvironmentResponseBody:
    title: TargetEnvironmentResponseBody
    type: object
    properties:
      apiKey:
        type: string
        description: The target environment
        example: Dolor ad non deleniti dolores blanditiis et.
    example:
      apiKey: Ut optio et atque eos ducimus expedita.
  UserGetUserBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Bad request, e.g. invalid data was sent in the request. (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetUserDetailsBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Bad request, e.g. invalid data was sent in the request. (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetUserDetailsInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Internal error. Check log for information. (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetUserDetailsResponseBody:
    title: 'Mediatype identifier: apiuserresult; view=default'
    type: object
    properties:
      paymentServerUrl:
        $ref: '#/definitions/PaymentServerUrlResponseBody'
      providerCallbackHost:
        type: string
        description: The provider callback host
        example: Voluptatibus autem ut ea adipisci perferendis.
      targetEnvironment:
        $ref: '#/definitions/TargetEnvironmentResponseBody'
    description: GetUserDetailsResponseBody result type (default view)
    example:
      paymentServerUrl:
        apiKey: Quidem illum voluptatibus dolor cum commodi ea.
      providerCallbackHost: Eveniet est voluptatum itaque dolor sunt.
      targetEnvironment:
        apiKey: Natus est molestiae quibusdam quis et.
  UserGetUserInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Internal error. Check log for information. (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetUserNotFoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Not found, reference id not found or closed in sandbox (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserGetUserRequestBody:
    title: UserGetUserRequestBody
    type: object
    properties:
      X-Reference-Id:
        type: string
        description: Format - UUID.
        example: Mollitia omnis.
    example:
      X-Reference-Id: Ut illo blanditiis veritatis ea qui aut.
    required:
    - X-Reference-Id
...
  UserNewUserBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Bad request, e.g. invalid data was sent in the request. (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserNewUserConflictResponseBody:
    title: 'Mediatype identifier: errorreason; view=default'
    type: object
    properties:
      code:
        type: string
        description: Code
        example: PAYER_NOT_FOUND
        enum:
        - PAYEE_NOT_FOUND
        - PAYER_NOT_FOUND
        - NOT_ALLOWED
        - NOT_ALLOWED_TARGET_ENVIRONMENT
        - INVALID_CALLBACK_URL_HOST
        - INVALID_CURRENCY
        - SERVICE_UNAVAILABLE
        - INTERNAL_PROCESSING_ERROR
        - NOT_ENOUGH_FUNDS
        - PAYER_LIMIT_REACHED
        - PAYEE_NOT_ALLOWED_TO_RECEIVE
        - PAYMENT_NOT_APPROVED
        - RESOURCE_NOT_FOUND
        - APPROVAL_REJECTED
        - EXPIRED
        - TRANSACTION_CANCELED
        - RESOURCE_ALREADY_EXIST
      message:
        type: string
        description: message
        example: Payee does not exist
    description: Conflict, duplicated reference id (default view)
    example:
      code: PAYER_NOT_FOUND
      message: Payee does not exist
  UserNewUserInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Internal error. Check log for information. (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserNewUserRequestBody:
    title: UserNewUserRequestBody
    type: object
    properties:
      providerCallbackHost:
        type: string
        description: The provider callback host
        example: Dolorem architecto aperiam deserunt.
    example:
      providerCallbackHost: Eos ut iusto nisi aliquam soluta.
...